pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "indusnet"
version = "0.1.0"
authors = [
  { name="KanroMiku", email="kanromiku_lpc@yeah.net" },
]
description = "一个用于工业互联网项目数据采集和处理的后端服务，仅作为课程作业使用"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "psycopg2-binary",
    "pika",
    "click",
    "pytest",
    "colorlog",
    "asyncpg",
    "aio_pika",
]

[project.scripts]
datalink = "cli:cli"

